Join vm and mk in a single program
This commit creates a wrapper for vm and mk; this required three
actions: create a single script that wraps "vm/mk" commands, move old
scripts to a new directory name src, and create a script for
install/uninstall. Fist, file easy-kernel-workflow.sh call all commands
available in vm and mk. Second, vm/mk have to be moved to a new
directory; as a result, some few changes were made in the file. Finally,
the "setup.sh" install the easy-kernel-workflow as a command in the home
directory and make it available in the path.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>